Combined genre file into app logic
